+++ tools/psr/psrctl/pkg/scenario/constants.go
@@ -10,8 +10,8 @@ const (
 	// LabelScenario has a bool value to indicate a resource is part of a scenario
 	LabelScenario = "psr.verrazzano.io/scenario"
 
-	// LabelScenarioId has a string value with the scenario ID
-	LabelScenarioId = "psr.verrazzano.io/scenario-id"
+	// LabelScenarioID has a string value with the scenario ID
+	LabelScenarioID = "psr.verrazzano.io/scenario-id"
 
 	// DataScenarioKey is the configmap key for the data field which
 	// has a base64 encoded Scenario
